feat(payment): PAYMENTS-6805 rename network error to server error, tighten logic
