+++ src/payment/strategies/ppsdk/step-handler/server-error.ts
@@ -1,8 +1,8 @@
 import { RequestError } from '../../../../common/error/errors';
 import { PaymentsAPIResponse } from '../ppsdk-payments-api-response';
 
-export const isNetworkError = (status: number): boolean =>
-    status >= 400;
+export const isServerError = (status: number): boolean =>
+    status < 200 || status >= 400;
 
-export const handleNetworkError = (response: PaymentsAPIResponse) =>
+export const handleServerError = (response: PaymentsAPIResponse) =>
     Promise.reject(new RequestError(response));
